Change Sequence<Byte> to Sequence<Integer> to fix tests

+++ cml-tool/src/main/java/org/fir3/cml/tool/util/seq/InputStreamSequence.java
@@ -8,7 +8,7 @@ import java.util.Objects;
  * A {@link Sequence} implementation that simply wraps an {@link InputStream}
  * instance.
  */
-public final class InputStreamSequence extends AbstractSequence<Byte> {
+public final class InputStreamSequence extends AbstractSequence<Integer> {
     private final InputStream source;
 
     /**
@@ -27,14 +27,14 @@ public final class InputStreamSequence extends AbstractSequence<Byte> {
     }
 
     @Override
-    protected Byte read0() throws IOException {
+    protected Integer read0() throws IOException {
         int nextByte = this.source.read();
 
         if (nextByte < 0) {
             return -1;
         }
 
-        return (byte) nextByte;
+        return nextByte;
     }
 
     @Override
